+++ backend/internal/model/permission/l10n/l10n.go
@@ -6,10 +6,10 @@ import (
 
 const Category = "permissions"
 
-func Name(permission int) string {
+func Name(permission uint8) string {
 	return fmt.Sprintf("%s:name%d", Category, permission)
 }
 
-func Desc(permission int) string {
+func Desc(permission uint8) string {
 	return fmt.Sprintf("%s:desc%d", Category, permission)
 }
